requirements.txt
flask
mysql-connector-python
flask-mail